Moved createViewSlice into features directory
