+++ client/src/components/view/DeleteButton.tsx
@@ -1,6 +1,6 @@
 import { IconButton } from '@mui/material';
 import { useAppDispatch } from '../../app/hooks';
-import { deleteNode, Id } from './createViewSlice';
+import { deleteNode, Id } from '../../features/createViewSlice';
 import { pinkColor } from './colors';
 import RemoveIcon from '@mui/icons-material/Remove';
 import { useStateKey } from './useStateKey';
